+++ src/AceSegmentWriter.h
@@ -43,8 +43,8 @@ SOFTWARE.
 #endif
 
 // Version format: xxyyzz == "xx.yy.zz"
-#define ACE_SEGMENT_WRITER_VERSION 100
-#define ACE_SEGMENT_WRITER_VERSION_STRING "0.1.0"
+#define ACE_SEGMENT_WRITER_VERSION 200
+#define ACE_SEGMENT_WRITER_VERSION_STRING "0.2.0"
 
 #include "ace_segment_writer/PatternWriter.h"
 #include "ace_segment_writer/NumberWriter.h"