+++ src/ace_segment_writer/NumberWriter.h
@@ -26,6 +26,7 @@ SOFTWARE.
 #define ACE_SEGMENT_WRITER_NUMBER_WRITER_H
 
 #include <stdint.h>
+#include <AceCommon.h> // decToBcd()
 #include "PatternWriter.h"
 
 namespace ace_segment {
@@ -92,10 +93,84 @@ class NumberWriter {
           pos, ((uint8_t) c < kNumHexCharPatterns) ? c : kHexCharSpace);
     }
 
-    /** Write the `len` hex characters given by `s` starting at `pos`. */
+    /**
+     * Write 2 hex characters given in the arguments. A convenience function
+     * that seems to occur frequently when writing time or date components.
+     */
+    void writeHexChars2At(uint8_t pos, hexchar_t c, hexchar_t d) {
+      writeHexCharAt(pos++ ,c);
+      writeHexCharAt(pos, d);
+    }
+
+    /** Write the `len` hex characters given by `s` at LED `pos`. */
     void writeHexCharsAt(uint8_t pos, const hexchar_t s[], uint8_t len) {
       for (uint8_t i = 0; i < len; ++i) {
         writeHexCharAt(pos++, s[i]);
+      }
+    }
+
+    /**
+     * Write a 2-digit BCD number at position, which involves just printing the
+     * number as a hexadecimal number. For example, 0x12 is printed as "12", but
+     * 0x1A is printed as "1 ".
+     */
+    void writeBcd2At(uint8_t pos, uint8_t bcd) {
+      uint8_t high = (bcd & 0xF0) >> 4;
+      uint8_t low = (bcd & 0x0F);
+      if (high > 9) high = kHexCharSpace;
+      if (low > 9) low = kHexCharSpace;
+      writeHexChars2At(pos, high, low);
+    }
+
+    /**
+     * Write a 2-digit decimal number at position digit, right justified
+     * with the given `padPattern` (default kPattern0). If the number is greater
+     * than 100, then print "  " (2 spaces). Useful for month, day, hour,
+     * minute, second.
+     *
+     * @param pos LED digit position
+     * @param d decimal value to write
+     * @param padPattern leading pattern, usually kPatternSpace or kPattern0
+     */
+    void writeDec2At(uint8_t pos, uint8_t d, uint8_t padPattern = kPattern0) {
+      if (d >= 100) {
+        mPatternWriter.writePatternAt(pos++, kPatternSpace);
+        mPatternWriter.writePatternAt(pos++, kPatternSpace);
+      } else {
+        uint8_t tens = d / 10;
+        uint8_t ones = d - 10 * tens;
+        if (tens == 0) {
+          mPatternWriter.writePatternAt(pos++, padPattern);
+        } else {
+          writeHexCharAt(pos++, tens);
+        }
+        writeHexCharAt(pos, ones);
+      }
+    }
+
+    /**
+     * Write the 4 digit decimal number `dd` at pos, right justified, padded
+     * with a the `padPattern` (default kPattern0). If the number is greater
+     * than 10000, print 4 spaces.
+     */
+    void writeDec4At(uint8_t pos, uint16_t dd, uint8_t padPattern = kPattern0) {
+      if (dd >= 10000) {
+        mPatternWriter.writePatternAt(pos++, kPatternSpace);
+        mPatternWriter.writePatternAt(pos++, kPatternSpace);
+        mPatternWriter.writePatternAt(pos++, kPatternSpace);
+        mPatternWriter.writePatternAt(pos++, kPatternSpace);
+      } else {
+        uint8_t high = dd / 100;
+        uint8_t low = dd - high * 100;
+        if (high == 0) {
+          mPatternWriter.writePatternAt(pos++, padPattern);
+          mPatternWriter.writePatternAt(pos++, padPattern);
+          writeDec2At(pos, low, padPattern);
+        } else {
+          writeDec2At(pos, high, padPattern);
+          pos += 2;
+          writeDec2At(pos, low, kPattern0);
+        }
       }
     }
 
@@ -105,6 +180,8 @@ class NumberWriter {
       b >>= 4;
       uint8_t high = (b & 0x0F);
 
+      // We can use writeInternalHexCharAt() because low and high are guaranteed
+      // to be a validate hex character.
       writeInternalHexCharAt(pos++, high);
       writeInternalHexCharAt(pos++, low);
     }
@@ -118,12 +195,14 @@ class NumberWriter {
     }
 
     /**
-     * Write the 16-bit unsigned number `num` as a decimal number at pos.
+     * Write the 16-bit unsigned number `num` as a decimal number at pos. The
+     * lead padding character is always kPatternSpace.
      *
      * @param pos start position of the number
      * @param num unsigned decimal number, 0-65535
      * @param boxSize size of box. This is meant to be similar to the "%-5d" or
-     *    "%5d" specifier to the printf() function
+     *    "%5d" specifier to the printf() function. The actual digits may spill
+     *    over the recommended box size.
      *    * 0 means no boxing, printing from left
      *    * > 0 means right justified inside box
      *    * < 0 means left justified inside box
@@ -166,26 +245,6 @@ class NumberWriter {
     }
 
     /**
-     * Write the 2 digit decimal number at pos. This method always writes 2
-     * characters. The padding is always a space, so 9 is printed as " 9" not
-     * "09". If num >= 100, then this prints "99".
-     *
-     * This method is meant to be a lighter version of writeUnsignedDecimalAt()
-     * when only 2 digits are needed. Seems to save about 110 bytes of flash.
-     */
-    void writeUnsignedDecimal2At(uint8_t pos, uint8_t num) {
-      if (num >= 100) {
-        writeHexCharAt(pos++, 9);
-        writeHexCharAt(pos++, 9);
-      } else {
-        uint8_t high = num / 10;
-        uint8_t low = num - high * 10;
-        writeHexCharAt(pos++, (high == 0) ? kHexCharSpace : high);
-        writeHexCharAt(pos++, low);
-      }
-    }
-
-    /**
      * Write the decimal point for the pos. Clock LED modules will attach the
      * colon segment to one of the decimal points.
      */
@@ -223,8 +282,9 @@ class NumberWriter {
     }
 
     /**
-     * Print the hex characters in `s` inside a box of size `boxSize` at
-     * position `pos`.
+     * Print the hex characters in `s` inside a recommended box of size
+     * `boxSize` at position `pos`. If the number of digits is bigger, the
+     * digits may spill over outside the box.
      *
      * @param boxSize if negative, left justified; if postive, right justified
      */
@@ -271,7 +331,7 @@ class NumberWriter {
      * @return index into buf that points to the start of the converted number,
      * e.g. for a single digit number, the returned value will be `bufSize-1`.
      */
-    uint8_t toDecimal(uint16_t num, hexchar_t buf[], uint8_t bufSize) {
+    static uint8_t toDecimal(uint16_t num, hexchar_t buf[], uint8_t bufSize) {
       uint8_t pos = bufSize;
       while (true) {
         if (num < 10) {